Move repositories and datasource inside dedicated packages
