+++ app/src/main/java/com/deezer/exoapplication/player/data/datasource/IsPlayingDataSource.kt
@@ -1,4 +1,4 @@
-package com.deezer.exoapplication.player.data
+package com.deezer.exoapplication.player.data.datasource
 
 import androidx.media3.common.Player
 import kotlinx.coroutines.channels.awaitClose
@@ -6,17 +6,15 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.callbackFlow
 import javax.inject.Inject
 
-class SongEndedDataSource @Inject constructor(
+class IsPlayingDataSource @Inject constructor(
     private val player: Player
 ) {
-
-    fun getSongEndedFlow(): Flow<Unit> = callbackFlow {
+    fun getIsPlayingFlow(): Flow<Boolean> = callbackFlow {
+        trySend(player.isPlaying)
         val listener = object: Player.Listener {
-            override fun onPlaybackStateChanged(state: Int) {
-                super.onPlaybackStateChanged(state)
-                if(state == Player.STATE_ENDED) {
-                    trySend(Unit)
-                }
+            override fun onIsPlayingChanged(isPlaying: Boolean) {
+                super.onIsPlayingChanged(isPlaying)
+                trySend(isPlaying)
             }
         }
         player.addListener(listener)
