+++ app/src/main/java/com/deezer/exoapplication/player/data/datasource/SongEndedDataSource.kt
@@ -1,4 +1,4 @@
-package com.deezer.exoapplication.player.data
+package com.deezer.exoapplication.player.data.datasource
 
 import androidx.media3.common.Player
 import kotlinx.coroutines.channels.awaitClose
@@ -6,15 +6,17 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.callbackFlow
 import javax.inject.Inject
 
-class IsPlayingDataSource @Inject constructor(
+class SongEndedDataSource @Inject constructor(
     private val player: Player
 ) {
-    fun getIsPlayingFlow(): Flow<Boolean> = callbackFlow {
-        trySend(player.isPlaying)
+
+    fun getSongEndedFlow(): Flow<Unit> = callbackFlow {
         val listener = object: Player.Listener {
-            override fun onIsPlayingChanged(isPlaying: Boolean) {
-                super.onIsPlayingChanged(isPlaying)
-                trySend(isPlaying)
+            override fun onPlaybackStateChanged(state: Int) {
+                super.onPlaybackStateChanged(state)
+                if(state == Player.STATE_ENDED) {
+                    trySend(Unit)
+                }
             }
         }
         player.addListener(listener)
