+++ app/src/main/java/com/deezer/exoapplication/player/data/repository/IsPlayingRepositoryImpl.kt
@@ -1,6 +1,7 @@
-package com.deezer.exoapplication.player.data
+package com.deezer.exoapplication.player.data.repository
 
-import com.deezer.exoapplication.player.domain.IsPlayingRepository
+import com.deezer.exoapplication.player.data.datasource.IsPlayingDataSource
+import com.deezer.exoapplication.player.domain.repository.IsPlayingRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
