+++ app/src/main/java/com/deezer/exoapplication/player/data/repository/SongEndedRepositoryImpl.kt
@@ -1,6 +1,7 @@
-package com.deezer.exoapplication.player.data
+package com.deezer.exoapplication.player.data.repository
 
-import com.deezer.exoapplication.player.domain.SongEndedRepository
+import com.deezer.exoapplication.player.data.datasource.SongEndedDataSource
+import com.deezer.exoapplication.player.domain.repository.SongEndedRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
