+++ app/src/main/java/com/deezer/exoapplication/player/di/PlayerModule.kt
@@ -1,12 +1,12 @@
 package com.deezer.exoapplication.player.di
 
-import com.deezer.exoapplication.player.data.InMemoryTrackRepository
-import com.deezer.exoapplication.player.domain.IsPlayingRepository
-import com.deezer.exoapplication.player.data.IsPlayingRepositoryImpl
-import com.deezer.exoapplication.player.data.SongEndedRepositoryImpl
-import com.deezer.exoapplication.player.domain.SongEndedRepository
+import com.deezer.exoapplication.player.data.repository.InMemoryTrackRepository
+import com.deezer.exoapplication.player.domain.repository.IsPlayingRepository
+import com.deezer.exoapplication.player.data.repository.IsPlayingRepositoryImpl
+import com.deezer.exoapplication.player.data.repository.SongEndedRepositoryImpl
+import com.deezer.exoapplication.player.domain.repository.SongEndedRepository
 import com.deezer.exoapplication.player.domain.QueueManager
-import com.deezer.exoapplication.player.domain.TrackRepository
+import com.deezer.exoapplication.player.domain.repository.TrackRepository
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
@@ -24,11 +24,11 @@ interface PlayerModule {
 
     @Singleton
     @Binds
-    fun bindSongRepository(impl: SongEndedRepositoryImpl): SongEndedRepository
+    fun bindSongEndedRepository(impl: SongEndedRepositoryImpl): SongEndedRepository
 
     @Singleton
     @Binds
-    fun bindSongRepository(impl: IsPlayingRepositoryImpl): IsPlayingRepository
+    fun bindIsPlayingRepository(impl: IsPlayingRepositoryImpl): IsPlayingRepository
 
     companion object {
         @Provides
