+++ app/src/main/java/com/deezer/exoapplication/player/domain/repository/TrackRepository.kt
@@ -1,4 +1,4 @@
-package com.deezer.exoapplication.player.domain
+package com.deezer.exoapplication.player.domain.repository
 
 import android.net.Uri
 import com.deezer.exoapplication.player.domain.model.Track
